+++ demos/cases/Filters.vue
@@ -48,6 +48,13 @@ const columns = [
 		options: countries,
 	},
 	{
+		id: "destinations",
+		header: { filter: { type: "multiselect" } },
+		footer: "Destinations",
+		options: countries,
+		width: 200,
+	},
+	{
 		id: "date",
 		header: { filter: "datepicker" },
 		template: obj => dateFormat.format(obj),