+++ demos/cases/InlineEditors.vue
@@ -41,8 +41,15 @@ const columns = [
 		header: 'Destinations - "multiselect"',
 		editor: {
 			type: "multiselect",
+			config: { clear: true },
 		},
 		options: countries,
+		width: 250,
+	},
+	{
+		id: "stars",
+		header: 'Stars - "text" (number)',
+		editor: { type: "text", config: { type: "number" } },
 		width: 250,
 	},
 ];