+++ src/components/Grid.vue
@@ -95,7 +95,6 @@ provide("grid-store", {
 	getReactiveState: dataStore.getReactive.bind(dataStore),
 	exec: firstInRoute.exec.bind(firstInRoute),
 	getRow: dataStore.getRow.bind(dataStore),
-	getRowIndex: dataStore.getRowIndex.bind(dataStore),
 });
 
 // auto config columns