+++ src/components/HeaderCell.vue
@@ -126,7 +126,7 @@ function getCell() {
 		:data-header-id="setID(column.id)"
 		:tabindex="!cell._hidden && column.sort && !cell.filter ? '0' : undefined"
 		role="columnheader"
-		:aria-colindex="cell._colindex"
+		:aria-colindex="column._colindex"
 		:aria-colspan="cell.colspan > 1 ? cell.colspan : undefined"
 		:aria-rowspan="cell.rowspan > 1 ? cell.rowspan : undefined"
 		:aria-sort="!sortMark?.order || cell.filter