+++ src/components/inlineFilters/filters.js
@@ -1,9 +1,11 @@
 import Text from "./Text.vue";
 import Richselect from "./Richselect.vue";
 import DatePicker from "./DatePicker.vue";
+import MultiSelect from "./MultiSelect.vue";
 
 export const filters = {
 	text: Text,
 	richselect: Richselect,
 	datepicker: DatePicker,
+	multiselect: MultiSelect,
 };